refactor(beaconevent): extract attribute helpers into attrs.go
Split the converter god file (1613 lines). Move the pure attribute/value utility
functions and the tool-name key-order lists into a new attrs.go in the same
package: AttrsToMap, MergeMaps, the First*/RunString/*Attr accessors, Int/Int64/
Float/Bool/StringSlice/AnyAttr, DecodeJSONValue, LegacyMessages, firstTextFromAny,
IsZeroJSON, Timestamp, Severity/SpanSeverity, and the toolNameKeys/genAIToolNameKeys/
mcpToolNameKeys/codexToolNameKeys vars.

Move-only, same package: behavior-preserving. converter.go drops to ~1224 lines;
exporter tests stay green.
